+++ packages/android_intent/android/src/main/java/io/flutter/plugins/androidintent/AndroidIntentPlugin.java
@@ -4,6 +4,7 @@
 
 package io.flutter.plugins.androidintent;
 
+import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
@@ -43,6 +44,10 @@ public class AndroidIntentPlugin implements MethodCallHandler {
         return Intent.ACTION_VOICE_COMMAND;
       case "settings":
         return Settings.ACTION_SETTINGS;
+      case "action_location_source_settings":
+        return Settings.ACTION_LOCATION_SOURCE_SETTINGS;
+      case "action_application_details_settings":
+        return Settings.ACTION_APPLICATION_DETAILS_SETTINGS;
       default:
         return action;
     }
@@ -123,6 +128,9 @@ public class AndroidIntentPlugin implements MethodCallHandler {
     if (mRegistrar.activity() == null) {
       intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
     }
+    if (call.argument("flag") != null) {
+      intent.addFlags((Integer) call.argument("flags"));
+    }
     if (call.argument("category") != null) {
       intent.addCategory((String) call.argument("category"));
     }
@@ -133,7 +141,12 @@ public class AndroidIntentPlugin implements MethodCallHandler {
       intent.putExtras(convertArguments((Map) call.argument("arguments")));
     }
     if (call.argument("package") != null) {
-      intent.setPackage((String) call.argument("package"));
+      String packageName = (String) call.argument("package");
+      intent.setPackage(packageName);
+      if (call.argument("componentName") != null) {
+        intent.setComponent(
+            new ComponentName(packageName, (String) call.argument("componentName")));
+      }
       if (intent.resolveActivity(context.getPackageManager()) == null) {
         Log.i(TAG, "Cannot resolve explicit intent - ignoring package");
         intent.setPackage(null);