+++ packages/android_intent/lib/android_intent.dart
@@ -14,37 +14,80 @@ const String kChannelName = 'plugins.flutter.io/android_intent';
 class AndroidIntent {
   /// Builds an Android intent with the following parameters
   /// [action] refers to the action parameter of the intent.
+  /// [flags] is the list of int that will be converted to native flags.
   /// [category] refers to the category of the intent, can be null.
   /// [data] refers to the string format of the URI that will be passed to
   /// intent.
   /// [arguments] is the map that will be converted into an extras bundle and
   /// passed to the intent.
+  /// [package] refers to the package parameter of the intent, can be null.
+  /// [componentName] refers to the component name of the intent, can be null.
+  /// If not null, then [package] but also be provided.
   const AndroidIntent({
     @required this.action,
+    this.flags,
     this.category,
     this.data,
     this.arguments,
     this.package,
+    this.componentName,
     Platform platform,
   })  : assert(action != null),
         _channel = const MethodChannel(kChannelName),
         _platform = platform ?? const LocalPlatform();
 
+  @visibleForTesting
+  AndroidIntent.private({
+    @required this.action,
+    @required Platform platform,
+    @required MethodChannel channel,
+    this.flags,
+    this.category,
+    this.data,
+    this.arguments,
+    this.package,
+    this.componentName,
+  })  : _channel = channel,
+        _platform = platform;
+
   final String action;
+  final List<int> flags;
   final String category;
   final String data;
   final Map<String, dynamic> arguments;
   final String package;
+  final String componentName;
   final MethodChannel _channel;
   final Platform _platform;
 
+  bool _isPowerOfTwo(int x) {
+    /* First x in the below expression is for the case when x is 0 */
+    return x != 0 && ((x & (x - 1)) == 0);
+  }
+
+  @visibleForTesting
+  int convertFlags(List<int> flags) {
+    int finalValue = 0;
+    for (int i = 0; i < flags.length; i++) {
+      if (!_isPowerOfTwo(flags[i])) {
+        throw ArgumentError.value(flags[i], 'flag\'s value must be power of 2');
+      }
+      finalValue |= flags[i];
+    }
+    return finalValue;
+  }
+
   /// Launch the intent.
   ///
-  /// This works only on Android platforms. Please guard the call so that your
-  /// iOS app does not crash. Checked mode will throw an assert exception.
+  /// This works only on Android platforms.
   Future<void> launch() async {
-    assert(_platform.isAndroid);
+    if (!_platform.isAndroid) {
+      return;
+    }
     final Map<String, dynamic> args = <String, dynamic>{'action': action};
+    if (flags != null) {
+      args['flags'] = convertFlags(flags);
+    }
     if (category != null) {
       args['category'] = category;
     }
@@ -56,10 +99,10 @@ class AndroidIntent {
     }
     if (package != null) {
       args['package'] = package;
+      if (componentName != null) {
+        args['componentName'] = componentName;
+      }
     }
-    // TODO(amirh): remove this on when the invokeMethod update makes it to stable Flutter.
-    // https://github.com/flutter/flutter/issues/26431
-    // ignore: strong_mode_implicit_dynamic_method
-    await _channel.invokeMethod('launch', args);
+    await _channel.invokeMethod<void>('launch', args);
   }
 }