+++ packages/google_sign_in/example/lib/main.dart
@@ -1,4 +1,4 @@
-// Copyright 2017 The Chromium Authors. All rights reserved.
+// Copyright 2019 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -54,7 +54,7 @@ class SignInDemoState extends State<SignInDemo> {
     });
     final http.Response response = await http.get(
       'https://people.googleapis.com/v1/people/me/connections'
-          '?requestMask.includeField=person.names',
+      '?requestMask.includeField=person.names',
       headers: await _currentUser.authHeaders,
     );
     if (response.statusCode != 200) {
@@ -115,11 +115,11 @@ class SignInDemoState extends State<SignInDemo> {
             leading: GoogleUserCircleAvatar(
               identity: _currentUser,
             ),
-            title: Text(_currentUser.displayName),
-            subtitle: Text(_currentUser.email),
+            title: Text(_currentUser.displayName ?? ''),
+            subtitle: Text(_currentUser.email ?? ''),
           ),
           const Text("Signed in successfully."),
-          Text(_contactText),
+          Text(_contactText ?? ''),
           RaisedButton(
             child: const Text('SIGN OUT'),
             onPressed: _handleSignOut,