+++ packages/path_provider/test/path_provider_test.dart
@@ -1,4 +1,4 @@
-// Copyright 2017 The Chromium Authors. All rights reserved.
+// Copyright 2019 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -7,8 +7,11 @@ import 'dart:io';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:path_provider/path_provider.dart';
+import 'package:platform/platform.dart';
 
 void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
   const MethodChannel channel =
       MethodChannel('plugins.flutter.io/path_provider');
   final List<MethodCall> log = <MethodCall>[];
@@ -17,6 +20,10 @@ void main() {
   channel.setMockMethodCallHandler((MethodCall methodCall) async {
     log.add(methodCall);
     return response;
+  });
+
+  setUp(() {
+    setMockPathProviderPlatform(FakePlatform(operatingSystem: 'android'));
   });
 
   tearDown(() {
@@ -33,6 +40,18 @@ void main() {
     expect(directory, isNull);
   });
 
+  test('getApplicationSupportDirectory test', () async {
+    response = null;
+    final Directory directory = await getApplicationSupportDirectory();
+    expect(
+      log,
+      <Matcher>[
+        isMethodCall('getApplicationSupportDirectory', arguments: null)
+      ],
+    );
+    expect(directory, isNull);
+  });
+
   test('getApplicationDocumentsDirectory test', () async {
     response = null;
     final Directory directory = await getApplicationDocumentsDirectory();
@@ -45,6 +64,50 @@ void main() {
     expect(directory, isNull);
   });
 
+  test('getApplicationSupportDirectory test', () async {
+    response = null;
+    final Directory directory = await getApplicationSupportDirectory();
+    expect(
+      log,
+      <Matcher>[
+        isMethodCall('getApplicationSupportDirectory', arguments: null)
+      ],
+    );
+    expect(directory, isNull);
+  });
+
+  test('getExternalStorageDirectory test', () async {
+    response = null;
+    final Directory directory = await getExternalStorageDirectory();
+    expect(
+      log,
+      <Matcher>[isMethodCall('getStorageDirectory', arguments: null)],
+    );
+    expect(directory, isNull);
+  });
+
+  test('getLibraryDirectory test', () async {
+    response = null;
+    final Directory directory = await getLibraryDirectory();
+    expect(
+      log,
+      <Matcher>[isMethodCall('getLibraryDirectory', arguments: null)],
+    );
+    expect(directory, isNull);
+  });
+
+  test('getExternalStorageDirectory iOS test', () async {
+    setMockPathProviderPlatform(FakePlatform(operatingSystem: 'ios'));
+
+    response = null;
+    try {
+      await getExternalStorageDirectory();
+      fail('should throw UnsupportedError');
+    } catch (e) {
+      expect(e, isUnsupportedError);
+    }
+  });
+
   test('TemporaryDirectory path test', () async {
     final String fakePath = "/foo/bar/baz";
     response = fakePath;
@@ -52,10 +115,38 @@ void main() {
     expect(directory.path, equals(fakePath));
   });
 
+  test('ApplicationSupportDirectory path test', () async {
+    final String fakePath = "/foo/bar/baz";
+    response = fakePath;
+    final Directory directory = await getApplicationSupportDirectory();
+    expect(directory.path, equals(fakePath));
+  });
+
   test('ApplicationDocumentsDirectory path test', () async {
     final String fakePath = "/foo/bar/baz";
     response = fakePath;
     final Directory directory = await getApplicationDocumentsDirectory();
+    expect(directory.path, equals(fakePath));
+  });
+
+  test('ApplicationSupportDirectory path test', () async {
+    final String fakePath = "/foo/bar/baz";
+    response = fakePath;
+    final Directory directory = await getApplicationSupportDirectory();
+    expect(directory.path, equals(fakePath));
+  });
+
+  test('ApplicationLibraryDirectory path test', () async {
+    final String fakePath = "/foo/bar/baz";
+    response = fakePath;
+    final Directory directory = await getLibraryDirectory();
+    expect(directory.path, equals(fakePath));
+  });
+
+  test('ExternalStorageDirectory path test', () async {
+    final String fakePath = "/foo/bar/baz";
+    response = fakePath;
+    final Directory directory = await getExternalStorageDirectory();
     expect(directory.path, equals(fakePath));
   });
 }