+++ packages/shared_preferences/test/shared_preferences_test.dart
@@ -7,6 +7,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
   group('$SharedPreferences', () {
     const MethodChannel channel = MethodChannel(
       'plugins.flutter.io/shared_preferences',
@@ -125,6 +127,15 @@ void main() {
           ));
     });
 
+    test('containsKey', () async {
+      const String key = 'testKey';
+
+      expect(false, preferences.containsKey(key));
+
+      preferences.setString(key, 'test');
+      expect(true, preferences.containsKey(key));
+    });
+
     test('clearing', () async {
       await preferences.clear();
       expect(preferences.getString('String'), null);
@@ -135,16 +146,55 @@ void main() {
       expect(log, <Matcher>[isMethodCall('clear', arguments: null)]);
     });
 
-    test('mocking', () async {
-      // TODO(amirh): remove this on when the invokeMethod update makes it to stable Flutter.
-      // https://github.com/flutter/flutter/issues/26431
-      // ignore: strong_mode_implicit_dynamic_method
-      expect(await channel.invokeMethod('getAll'), kTestValues);
+    test('reloading', () async {
+      await preferences.setString('String', kTestValues['flutter.String']);
+      expect(preferences.getString('String'), kTestValues['flutter.String']);
+
       SharedPreferences.setMockInitialValues(kTestValues2);
-      // TODO(amirh): remove this on when the invokeMethod update makes it to stable Flutter.
-      // https://github.com/flutter/flutter/issues/26431
-      // ignore: strong_mode_implicit_dynamic_method
-      expect(await channel.invokeMethod('getAll'), kTestValues2);
+      expect(preferences.getString('String'), kTestValues['flutter.String']);
+
+      await preferences.reload();
+      expect(preferences.getString('String'), kTestValues2['flutter.String']);
+    });
+
+    test('back to back calls should return same instance.', () async {
+      final Future<SharedPreferences> first = SharedPreferences.getInstance();
+      final Future<SharedPreferences> second = SharedPreferences.getInstance();
+      expect(await first, await second);
+    });
+
+    group('mocking', () {
+      const String _key = 'dummy';
+      const String _prefixedKey = 'flutter.' + _key;
+
+      test('test 1', () async {
+        SharedPreferences.setMockInitialValues(
+            <String, dynamic>{_prefixedKey: 'my string'});
+        final SharedPreferences prefs = await SharedPreferences.getInstance();
+        final String value = prefs.getString(_key);
+        expect(value, 'my string');
+      });
+
+      test('test 2', () async {
+        SharedPreferences.setMockInitialValues(
+            <String, dynamic>{_prefixedKey: 'my other string'});
+        final SharedPreferences prefs = await SharedPreferences.getInstance();
+        final String value = prefs.getString(_key);
+        expect(value, 'my other string');
+      });
+    });
+
+    test('writing copy of strings list', () async {
+      final List<String> myList = <String>[];
+      await preferences.setStringList("myList", myList);
+      myList.add("foobar");
+
+      final List<String> cachedList = preferences.getStringList('myList');
+      expect(cachedList, <String>[]);
+
+      cachedList.add("foobar2");
+
+      expect(preferences.getStringList('myList'), <String>[]);
     });
   });
 }