+++ packages/webview_flutter/android/src/main/java/io/flutter/plugins/webviewflutter/WebViewFlutterPlugin.java
@@ -8,13 +8,13 @@ import io.flutter.plugin.common.PluginRegistry.Registrar;
 
 /** WebViewFlutterPlugin */
 public class WebViewFlutterPlugin {
-
   /** Plugin registration. */
   public static void registerWith(Registrar registrar) {
     registrar
         .platformViewRegistry()
         .registerViewFactory(
-            "plugins.flutter.io/webview", new WebViewFactory(registrar.messenger()));
+            "plugins.flutter.io/webview",
+            new WebViewFactory(registrar.messenger(), registrar.view()));
     FlutterCookieManager.registerWith(registrar.messenger());
   }
 }